refactor(abstract-utxo): namespace fixedScript methods
We will add descriptor-focused methods soon so let's clean this up

Issue: BTC-1450
