+++ modules/abstract-utxo/src/transaction/fixedScript/index.ts
@@ -1,0 +1,1 @@
+export { explainPsbt, explainLegacyTx } from './explainTransaction';
